Handle resemble output setting applying to every image
resemble.outputSettings() applied the settings on a global scale causing it to affect every diff image
on the current page. This fix is directed to resolve that issue by implementing the output settings right
before the diffProcess is created thus applying the settings to its respective diff image with its own controls.

Change-Id: I69def795b44158a1771df0582c4711ad1bae7af7