Update resemble Resemble 3.2.4 > 4.0.0
Resemble.JS 3.2.4 has support for rendering with Node.JS and as such
depends on canvas, which depends on cairo and in turn requires system
libraries to be available. The missing libraries cause the build to
fail.

Since this is a browser plugin, we do not need the canvas dependency. I
tried to find a way to use `yarn install --ignore-optional`, but short
of editing the Gerrit main WORKSPACE, I could not find anyway to do it.

Eventually I have found upstream changed canvas to be an optional
dependencies:
https://github.com/rsmbl/Resemble.js/commit/35e294dab95ad6bb531138eb5613355571805aa9
It was released with version 4.0.0 of Resemble.JS and fixes the build
for me.

I have noticed the `yarn.lock` is still pointing to resemble@1.2.1 which
clearly got forgotten. I don't know how the build work since Gerrit
`yarn_install` refers to that lock file. I have regenerated it with
Gerrit 3.5 using:

bazel run @yarn//:yarn install -- --cwd plugins/image-diff install

The `package-lock.json` has been regenerated with npm 7.21.0 within the
Gerrit stable-3.5 branch (for polymer-bridges).

Change-Id: I924b66b535619a4072039b7396850ca34caf9e3e