Add .zuul.yaml
Adds a .zuul.yaml to setup a job for image-diff.
This allows us to customise things without having
to do this in the zuul/config repo.
We also change what artifact we save and also
don’t run bazel tests as we don’t have it defined.
Doing that fixes the job, hence why we are doing
it in this repo as it’s a custom job.

Change-Id: If9af4729fce7529d7d972bc83a4a25e9f8ade802